# openapi.yaml
openapi: 3.1.0
info:
  title: Meraki License Server API
  version: 1.0.0
  description: |
    REST API for managing restaurant-CMS licenses.
    - **Public** endpoints (`/api/v1/`) — no auth, used by CMS instances
    - **Portal** endpoints (`/api/portal/`) — JWT Bearer (customer session)
    - **Admin** endpoints (`/api/admin/`) — JWT Bearer (admin session)
servers:
  - url: /api/v1
    description: Public API
  - url: /api/portal
    description: Customer Portal API
  - url: /api/admin
    description: Admin API
...
components:
  securitySchemes:
    portalAuth:
      type: http
      scheme: bearer
      description: Customer portal JWT (from POST /api/portal/login)
    adminAuth:
      type: http
      scheme: bearer
      description: Admin JWT (from POST /api/admin/login)
    licenseAuth:
      type: http
      scheme: bearer
      description: Signed license JWT (from POST /api/v1/validate)
...
paths:
...
  /admin/customers/gdpr-requests:
    get:
      tags: [admin]
      summary: List pending deletion requests (SuperAdmin)
      security: [adminAuth: []]
      responses:
        '200': { description: Deletion request list }
  /admin/customers/gdpr-requests/{id}/process:
    post:
      tags: [admin]
      summary: Anonymize customer data (SuperAdmin)
      security: [adminAuth: []]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        '200': { description: Customer anonymized, invoices retained }
...